concepts/pptx-hub/pyproject.toml
[project]
name = "pptx-hub"
version = "0.1.0"
description = "Production-ready PowerPoint processing platform with multi-tenancy support"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = [
    "powerpoint", "pptx", "translation", "django", 
    "multi-tenant", "document-processing", "presentation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=1.0.0",
    "pydantic>=2.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
django = [
    "django>=5.0",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.27.0",
    "django-filter>=24.0",
    "django-q2>=1.6.0",
]
storage = [
    "boto3>=1.34.0",
]
translation = [
    "deepl>=1.18.0",
]
cli = [
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.8.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]
all = [
    "pptx-hub[django,storage,translation,cli]",
]

[project.scripts]
pptx-hub = "pptx_hub.cli.main:app"

[project.urls]
Homepage = "https://github.com/YOUR_ORG/pptx-hub"
Documentation = "https://pptx-hub.readthedocs.io"
Repository = "https://github.com/YOUR_ORG/pptx-hub"
Changelog = "https://github.com/YOUR_ORG/pptx-hub/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pptx_hub"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "TCH", "PTH", "RUF"]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["pptx_hub"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["pptx.*", "django.*", "rest_framework.*", "boto3.*", "deepl.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-ra -q --cov=pptx_hub --cov-report=term-missing"

[tool.coverage.run]
source = ["src/pptx_hub"]
branch = true
